requirements.txt
streamlit>=1.53.0
requests>=2.28.0
python-dotenv>=0.19.0
langchain>=0.1.0
langgraph>=0.0.1
openai>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
psycopg2-binary>=2.9.0
